# Speed-up flame graph generation by skipping unneeded work.